stocksbackend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
cryptography==42.0.5
djangorestframework-simplejwt==5.3.1
pycparser==2.22
pyjwt==2.8.0
pytest-django==4.8.0
celery==5.4.0
redis==5.0.6
requests==2.32.3
pandas==2.2.2
numpy==2.0.0
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-rest-passwordreset==1.4.1
faker==26.1.0